pyproject.toml
[project]
name = "rl-engine"
version = "0.1.0"
description = "Ralph Loop Bug Fix Engine — agentic SDLC for GitHub organizations"
requires-python = ">=3.12"
dependencies = [
    "google-genai",
    "anthropic",
    "httpx",
    "pyyaml",
    "jinja2",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[project.scripts]
rl-engine = "engine.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
